Made fixes due to Java 8 and UIMA 2.9.0.
UIMA 2.9.0 has sometimes different orderings of XMI elements which
effected a test in JNET.
Also often the default element order in some collections is different in
Java8 which also caused failing tests.
Changed the JNET ConsistencyPreservation to really account for all
entities and not just for the first (whichever that is) of an entity
string.