New interface, new index.
Now all annotation indexes implement the JCoReAnnotationIndex interface.
This is mainly thought as giving some central point gathering all
annotation indexes for an overview. An issue is that it might not be
clear when which index is used best.
Added an index for to retrieve annotations covered by some given span.