Fixed tests of the ConllConsumer.
In one test, an NPE was thrown because lemmas were activated but there
were no lemmas in the test data. So lemmas were deactivated via the
respective parameter at consumer creation.
The other test tested for the output file 0.CONLL. This file is
generated when the CAS had no Header or the Header does not specify an
ID. But the test specified an ID, so the file was never written.
Commented out the test looking for the file.