Merging the XML multiplier code and descriptors into the XML reader.
The idea is that the reader, the multiplier and the multiplierreader are quite related and also share a lot of code. Thus we will try to have all three components in one project.
Also did some renaming, away from pubmed specific stuff to just "XML" in the base project. For a PubMed specific component, a JCoRe project should be created.