Adding the URI type for CAS multipliers.
When using CAS multipliers, the reader sends a CAS that only contains a reference to the resource that should actually be read. The type system of this CAS should be as minimal as possible for transport over the wire. Thus, the multiplier type systems should only contain the absolutely necessary information.