Added a configurable external resource definition to a new gazetteer descriptor.
By using the configurableDataResourceSpecifier element, the
class implementation for ResourceSpecifier of the same name is used.
This one has meta data and, thus, a configuration. So we don't need
to rely on the configuration file. This help especially for the
PipelineBuilder because then, all parameters are well defined and
can be edited.