Handled a rare error with acronym resolution but did no solve it.
When the input text containes newlines it might happen that an
abbreviation long form has the newline in it. Actually, this was only
observed with faulty documents (PMID 26306035 has some weird abstract
section causing the issue) or sentence splitting errors. We now catch
the issue, do not create a new Mention and continue.