Added security checks if actually data from the database was retrieved.
When a lot of pipelines are started on a small table it can
actually happen that all pipelines find the unprocessed table but
when they try to get it, nothing is left. Then, the inherited
data retrieval method from the (Abstract)DbReader returns null
which lead to an NPE in one case. This should now be fixed.