The PMC URI search now happens concurrently.
That means that an extra thread is started to search for the files
to be read. The reader thread can get the URIs as they become
available and so, processing can already start while the files
are still assembled. The reason for this change is that reading
all of the PMC file names (not loading the files themselves or even
parsing them) took around 30 minutes. Now, the parsing can begin
almost instantaneously.