Fixed an issue where a ZIP entry was allegedly not existent.
An entry with a '/' included in the name there were difficulties.
Those are now resolved. Also we changed from the ZipFileSystem to
a plain ZipFile. The ZipFile entries don't give as a URI reference
directly so we have to build it ourselves. However, reading the
list of entries is a magnitude faster than for the ZipFileSystem.
The file system had to be searched as a real file system instead of just
using the Zip-internal table of content.