Lowered an error log message to debug level because we ignore it anyway.
In BioSem there are relatively regular errors where the code of
BioSem loses its integrity. This is mostly due to the fact that
it works quite a lot with String.indexOf instead of actually using
token offsets. Some of these issue had been fixed in the past, others
didn't seem to be fixed easily and so they were just accepted. Log
messaged indicate those errors, though. But when they are ERROR level
log messages, they render the search for "actual" errors very
difficult. Since we ignore the BioSem errors and accept them as
shortcomings of the algorithm, the log level is lowered to debug.