Enhanced the map providers for more file format variants and persistence.
The reading logic has been shifted into the new AbstractMapProvider class.
The new MapProvider class is basically the old class but extends from
the AbstractMapProvider. The map file delimiter is now automatically
recognized and can be either a tab or whitespace.
New class "ReversedDoubleMapProvider" has been added to read a file
that has double values in the left and the actual map keys in the
right columns.
the MapDBReversedDoubleMapProvider is untested. It uses the MapDB
library to create a persistent map (in a temporary file, however)
in order to save memory. It was not required in the end but might
still work fine. Note, however, that the MapDB dependency is set
to 'provided', that is, in a project that wants to use this provider,
the MapDB dependency must be added manually.