Setting the default batch size of the DBReader from 50 to 1000.
The observation is that the search for unprocessed documents is
quite an overhead. But once the table position has been found, where
documents are still unprocessed, returning a larger batch is quick.
As a consequence, the database can deliver a lot more documents
per second with a higher batch size, also resulting in a better
CPU usage of the pipeline.