Reactivating test multithreading, I think I found the issue.
The flair components start their tests with the default "python"
command line which is not the correct one in the travis VM.
The in such cases, when the external program failed, the STDIO bridge
would not terminate at all. There is now an "error termination" signal.
In this commit, the original cause for the hanging is not yet fixed,
I want to check that instead the error is recognized and the bridge
terminates.
I did not notice this cause because of the multithreading where
the hanging wasn't obvious because other tests just continued
and "scolled away" the hangig part.