Fixed an issue where FLAIR used slightly other tokens that UIMA.
The cause is that in rare cases a token produced by JTBD may
contain a whitespace. For FLAIR, all white spaces are token boundaries.
For the conversion from FLAIR offsets to UIMA offsets, tokens
containing whitespaces are now interpreted as two tokens. The actual
token annotation of the document is not not changed, however.