FLAIR-NER: All tests running. Preliminary completion of the current work.
The `jcore-flair-ner-ae` is now able to return the contextualized
word embeddings created for sequence tagging. Those are added to the
`Token.embeddingVectors` feature and can thus be used in downstream
tasks.
Note that the tests will currently not run due to the fact that
the required flair version is a nightly build that allows to read
the word embeddings after sequence tagging. This will be fixed
by the next flair release or the use of the nightly build in travis.