Refactored the XMIDBWriter to find missing binary mapping items batch-wise.
Previously, each document had its missing mapping items computed
individually and the database table was updated for each single
document as well. Apart from an abundance of database access, this
proved to be a major bottleneck if the mapping grows large as it does
for Lemmas, for example, which might still be worth mapping (further
experiments pending here). Each new mapping item requires an
exclusive lock on the mapping table which slows things down.
We now update batch-wise. This will hopefully speed things up a lot.