Re-created the waiting-for-required-mappings mechanism.
Now, each thread checks if all mappings required are already there.
If so, it creates a list of required items which can be processed
by any other thread.
It then tries to get the lock on the mapping update. If it gets it,
it processes its own and also the lists of required items from other threads.
Otherwise, it waits for another thread to process its items.