Handling document text hash mismatches in the AnnotationAdder.
There is a new annotation format, `SimpleTSVEntityWithDocumentTextShaAnnotationFormat.java`
which expects the base64-encoded document sha256 hash from the original document text on
which an annotation was made. If it is given, the
`TextAnnotationListAdder` compares this with the base64 encoded sha256
digest from the current CAS document text. On a mismatch, the annotation
won't be added to avoid offset errors later on. A warning will be issued.
If the new parameter `PreventProcessedMarkOnDigestMismatch` is `true`,
the adder will also try to find the `DBProcessingMetaData` annotation
in the current CAS and set it to prevent the document from being
marked as processed in the JeDIS subset table by the `DBCheckpointAE`.
This makes only sense, of course, if the current pipeline is reading
from such a table. An error will be logged if the `DBProcessingMetaData`
annotation cannot be found.