Adding the `DocumentReleaseCheckpoint` to the `DBCheckpointAE`.
This facility is meant to help only mark those document IDs as `processed`
that have been really finished processing in all components and
all information has been written to file or the database. All components
writing information into the database in a batch-wise manner should
use this class to register themselves in `initialize` and then release
the `DocumentID`s of the documents for which they have sent all information
to the database and are really finished with those documents. Only then
the `DBCheckpointAE` will mark the documents as being finished with
processing.