Enhancing the JeDIS synchronization.
Actually, trying to fix it. The last approach used the comonent
instances as synchronization keys which fails in a multithreaded
environment because  no document is processed by different threads.
Now, the XMI DB Writer and the Checkpoint AE take a new
synchronization key parameter which is used instead. This is the
most general solution: It also works in the case that multiple DB Writers
should be in the pipeline. They, each writer only needs to get another
synchronization key. But multi-threaded versions of the same writer
will have the same key which is what we want.