Created code to fetch the hashes of existing XMI documents in the XMI table.
However, the code is currently in the wrong place. It must go the XML reader. We need to compute the hash directly after parsing the document text from XML so we can then compare it to the hashes in the database. When we would use the document text in the XMI reader this would always be the same as the database hash because we have read the document from the very same database table.
So, next up is code movement to the XML reader.