JTBD: Fixed an issue where a regular expression match took days without finishing.
The issue occurred with document PMC7575323. The respective regex, `(.*[\W].*){5,}`, has an unbounded number of matches allowed. In a first try, I introduced an upper bound of 50 but this did still take several minutes without finishing. Instead, the expression is now only applied for "superunits" (I'm actually not sure what those consist of) with a length of at most 200 characters. With that change, the document in question was finished within seconds.