Adding another repository with https protocol in context of BioLemmatizer.
BioLemmatizer uses MorphAdorner and has a repository for it in its own pom.xml. However, it again has no https and is blocked by Maven. I try to add the repository in the jcore-biolemmatizer-ae pom.xml in the hope that this repository will then be used instead of the original one.