ES Consumer: Adding a Lucene index cache for the Map- and AddonProviders.
When the mapping files get very large (in my case over 30M lines) loading all of it into memory can cause memory issues. Those can be solved frequently by using internalization where all Strings are uniquely stored in the String pool. However, this pool is also of a finite size (which can be set via the -XX:StringTableSize=<size> JVM option). In the case that most probably only a part of the whole map will ever be used, an alternative way are the new, abstract `PersistentIndexAddonTermsProvider` and `PersistentStringIndexMapProvider` classes. Both have one concrete sub class which uses a Lucene index to store the files in an indexed format. A Guava cache is used to load from this index and keep the actually required elements in memory. This relieved memory requirements a lot when dealing with very large resource files. Resolves #123.