DocumentReleaseCheckpoint: Fixed a bug where documents were not marked as being processed.
The issue was that we counted how often a document was released and compared that number to the number of registered components. However, when a component de-registered itself but had released its documents before, their count was actually too high. The remedy - without introducing other sources of error - was to not just count the number of releases per document but to explicitly track the componentIds that had released each documents.