Add fallbacks in case of unknown XML tag sets for PMC reader.
Newer JATS versions are - until now - backward compatible with their previous version. Since we don't use any of the special capabilities of any JATS format, we just check for JATS in general and assign the latest version if we haven't handled the exact version.
We do the same for the NLM tagsets.