Add a mechanism to avoid mirror subset reset for updated JeDIS document whose text hasn't changed; resolves #130.
The XMLDBMultiplier already had the "ToVisit" mechanism for the AnnotationDefinedFlowController that could skip the majority of a pipeline when the document hash in the database was the same as the hash for the updated document text. However, we needed to update the base documents and, thus, needed to not skip the XMI writer. But by default, that would cause the mirror subsets to reset.
Now, the XMLDBMultiplier fills the new feature of the DBProcessingMetaData annotation named "isDocumentHashUnchanged". If it is set to be unchanged, the XMIDBWriter will not reset the mirror subsets for that document.