Adding a feature to first delete index documents associated with the current CAS document and then index anew.
In relation index there is the issue that depending on some non-deterministic annotation orderings the relation annotation IDs differ from pipeline run to pipeline run. This causes the appearance of index document duplicates just because their ID is different due to some side effects. The new feature should allow to first delete all index documents associated with the PubMed document to avoid this issue.
Currently working on a test for this.