Add a function to delete index documents before sending the current documents to the index.
This is useful when the index documents have IDs that are not the document ID itself. For example, when indexing entities, the index document IDs will be something like the document ID plus the entity index in the document. If those entity index documents have a field disclosing the original document ID, this field can be used to delete all index documents associated with that document first. If then the entity tagger is changed, for example, old, deprecated documents can be removed instead of staying in the index, e.g. when the new tagger recognizes less entities than the old one.