Add the `next-concept` assignment strategy.
Christine Engelmann, who had developed the likelihood scale, the likelihood dictionary and this component, told me that the likelihood assignment strategy used in this component performed a bit worse that a simple 'assign likelihood to the direct next concept' in her evaluations. Thus, we added the alternate strategy. The strategy to use is set via a component parameter.